Move Code from App.tsx to StopWatch to make tests pass
